lmkd: don't close trace_marker fd when EBADF is returned
Kernel returns EBADF when tracing_on = 0 in which case we will close and
reopen trace_marker for next write which can add overhead because of
system checks that open() system call will undergo. As this is not a
real Bad file descriptor, choose not to close the trace_marker fd.

CRs-Fixed: 2611369
Change-Id: Id1bd48f11329dd3471ea2a49c774e9883b6a04c0