lmkd: fix polling logic of super critical events
For lmkd old strategy, it is possible that the polling logic stuck in
while(1) for a super critical event because of poll_now = 'true' and
later call_handler() is skipped, thus the poll_params.poll_handler will
be not NULL till the call_handler() is called and set it to NULL under
POLLING_DO_NOT_CHANGE when the polling exceeds the psi_window_size. Fix
it.

CRs-Fixed: 2984011
Change-Id: If7dda4fb3a3264607f5d22f6d2f889f646295f2b
Signed-off-by: Charan Teja Reddy <charante@codeaurora.org>