lmkd: Resolve Cts SequentialRW test failures
This test case causes the file cache to expand to cover all available
memory, with kswapd running continuously taking pages from the inactive
list. This was observed to generate kills due to either
PRESSURE_AFTER_KILL or CRITICAL_KILL. Limit both of these events to
only affect processes with adj >= 50, to exclude CTS which runs with
adj 0.

Increase the super critical event threshold to 400 ms, and handle it
the same behaviour as the old CRITICAL_KILL event. This 400ms
threshold was found experimentally to not be reached by this particular
CTS case.

Additionally, limit thrashing kills to tasks with ADJ scores of
VISIBLE_APP_ADJ.

CRs-Fixed: 3045668
Change-Id: I1a806607df32d95069a86babb95721d2ce7b5f2d